refactor: revamp the process logic
fix #189 #188
