+++ src/burn/render_then_merge.py
@@ -4,9 +4,9 @@ import argparse
 import os
 import subprocess
 from src.config import GPU_EXIST, SRC_DIR
-from src.burn.generate_danmakus import get_resolution, process_danmakus
-from src.burn.generate_subtitles import generate_subtitles
-from src.burn.render_video import render_video
+from src.danmaku.generate_danmakus import get_resolution, process_danmakus
+from src.subtitle.generate_subtitles import generate_subtitles
+from src.burn.render_command import render_command
 from src.upload.extract_video_info import get_video_info
 from src.log.logger import scan_log
 
@@ -20,18 +20,18 @@ def normalize_video_path(filepath):
     new_date_time = f"{date_time_parts[0][:4]}-{date_time_parts[0][4:6]}-{date_time_parts[0][6:8]}-{date_time_parts[1]}-{date_time_parts[2]}"
     return filepath.rsplit('/', 1)[0] + '/' + parts[0] + '_' + new_date_time + '-.mp4'
 
-def merge_videos(in_final_video, title, artist, date, merge_list):
+def merge_command(in_final_video, title, artist, date, merge_list):
     """Merge the video segments and preserve the first video's metadata
     Args:
         in_final_video: str, the path of videos will be merged
     """
-    merge_command = [
+    command = [
     'ffmpeg', '-f', 'concat', '-safe', '0', '-i', merge_list, '-metadata', f'title={title}', '-metadata', f'artist={artist}', '-metadata', f'date={date}', '-use_wallclock_as_timestamps', '1', 
     '-c', 'copy', in_final_video
     ]
     try:
         scan_log.info("Begin merging videos...")
-        result = subprocess.run(merge_command, check=True, capture_output=True, text=True)
+        result = subprocess.run(command, check=True, capture_output=True, text=True)
         scan_log.debug(f"FFmpeg output: {result.stdout}")
         if result.stderr:
             scan_log.debug(f"FFmpeg debug: {result.stderr}")
@@ -39,7 +39,7 @@ def merge_videos(in_final_video, title, artist, date, merge_list):
         scan_log.error(f"Error: {e.stderr}")
     subprocess.run(['rm', merge_list], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
 
-def render_and_merge(video_path_list):
+def render_then_merge(video_path_list):
     title = ''
     artist = ''
     date = ''
@@ -72,7 +72,7 @@ def render_and_merge(video_path_list):
             if GPU_EXIST:
                 generate_subtitles(original_video_path)
             # Burn danmaku or subtitles into the videos 
-            render_video(original_video_path, video_to_be_merged, subtitle_font_size, subtitle_margin_v)
+            render_command(original_video_path, video_to_be_merged, subtitle_font_size, subtitle_margin_v)
             if not os.path.exists(merge_list):
                 open(merge_list, 'w').close() 
             with open(merge_list, 'a') as f:
@@ -87,15 +87,8 @@ def render_and_merge(video_path_list):
         # test_path = original_video_path[:-4]
         # os.rename(original_video_path, test_path)
         
-    merge_videos(output_video_path, title, artist, date, merge_list)
+    merge_command(output_video_path, title, artist, date, merge_list)
     subprocess.run(['rm', '-r', tmp])
 
     with open(f"{SRC_DIR}/upload/uploadVideoQueue.txt", "a") as file:
         file.write(f"{output_video_path}\n")
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='Danmaku burns')
-    parser.add_argument('video_path', type=str, help='Path to the Video file')
-    args = parser.parse_args()
-    video_path = args.video_path
-    render_and_merge(video_path)
