+++ src/danmaku/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) 2024 bilive.
+
+import sys
+import os
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
